% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp1d_long_data}
\alias{.overimp1d_long_data}
\title{Build a long data frame of observed and imputed values for a variable}
\usage{
.overimp1d_long_data(obj, var.name, data, dataset = c("train", "test"))
}
\description{
Build a long data frame of observed and imputed values for a variable
}
\keyword{internal}
