% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp1d_subtitle}
\alias{.overimp1d_subtitle}
\title{Subtitle used across overimputation displays}
\usage{
.overimp1d_subtitle(m, var.name)
}
\description{
Subtitle used across overimputation displays
}
\keyword{internal}
