% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp2d_long_data}
\alias{.overimp2d_long_data}
\title{Build long-format data for two-variable overimputation plots}
\usage{
.overimp2d_long_data(obj, var.x, var.y, data, dataset = c("train", "test"))
}
\description{
Build long-format data for two-variable overimputation plots
}
\keyword{internal}
