% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp2d_subtitle}
\alias{.overimp2d_subtitle}
\title{Subtitle helper for two-variable displays}
\usage{
.overimp2d_subtitle(var.x, var.y)
}
\description{
Subtitle helper for two-variable displays
}
\keyword{internal}
