% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp_as_data_frame}
\alias{.overimp_as_data_frame}
\title{Coerce tabular inputs to plain data frames}
\usage{
.overimp_as_data_frame(data, label)
}
\description{
Coerce tabular inputs to plain data frames
}
\keyword{internal}
