% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.overimp_components}
\alias{.overimp_components}
\title{Retrieve the masking matrix and imputed datasets for a given split}
\usage{
.overimp_components(obj, dataset = c("train", "test"))
}
\description{
Retrieve the masking matrix and imputed datasets for a given split
}
\keyword{internal}
