% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_nhanes3.R
\docType{data}
\name{imp_nhanes3}
\alias{imp_nhanes3}
\title{Precomputed mixgb imputed datasets for `nhanes3`}
\format{
A list of 5 data.frames (each a completed dataset) created by
  `mixgb::mixgb()` in development.
}
\source{
Generated during package development with `mixgb::mixgb()`.
}
\usage{
data(imp_nhanes3)
}
\description{
A small precomputed list object containing 5 imputed datasets generated by
`mixgb::mixgb()` on the `nhanes3` example data. This dataset is included so
that users can run plotting examples without installing `mixgb`.
}
\keyword{datasets}
