% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_nhanes3_newborn.R
\docType{data}
\name{imp_nhanes3_newborn}
\alias{imp_nhanes3_newborn}
\title{Precomputed mixgb imputed datasets for `nhanes3_newborn`}
\format{
A list of `m` data.frames (each a completed dataset) created by
  `mixgb::mixgb()` in development.
}
\source{
Generated during package development with `mixgb::mixgb()`.
}
\usage{
data(imp_nhanes3_newborn)
}
\description{
A small precomputed list object containing `m` imputed datasets generated by
`mixgb::mixgb()` on the `nhanes3_newborn` example data. This dataset is included so
that users can run plotting examples without installing `mixgb`.
}
\examples{
data(imp_nhanes3_newborn)
str(imp_nhanes3_newborn)
}
\keyword{datasets}
