% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimp1D.R
\name{overimp1D_qq}
\alias{overimp1D_qq}
\title{plot qqplot for numeric}
\usage{
overimp1D_qq(
  obj,
  var.name,
  train.data,
  test.data = NULL,
  point.size = 1,
  ylim = NULL
)
}
\description{
plot qqplot for numeric
}
