% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimp1D.R
\name{overimp1D_qqline}
\alias{overimp1D_qqline}
\title{plot qqline plot for numeric}
\usage{
overimp1D_qqline(obj, var.name, train.data, test.data = NULL, ylim = NULL)
}
\description{
plot qqline plot for numeric
}
