% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimp_traintest_qq.R
\name{overimp_traintest_qq}
\alias{overimp_traintest_qq}
\title{plot qqplot for overimputation of training set and test set}
\usage{
overimp_traintest_qq(obj, var.name, train.data, test.data, color.pal = NULL)
}
\description{
plot qqplot for overimputation of training set and test set
}
