% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimpute.R
\name{overimpute}
\alias{overimpute}
\title{Overimpute main function}
\usage{
overimpute(train.data, test.data = NULL, p = 0.3, m = 5, method = "mixgb", ...)
}
\arguments{
\item{method}{can be one of the following: "mixgb","miae","mice","cart" or "ranger"}
}
\description{
Overimpute
}
