% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimpute_mice.R
\name{overimpute_mice}
\alias{overimpute_mice}
\title{Overimputation using mice default (can only do with training data)}
\usage{
overimpute_mice(
  train.data,
  test.data = NULL,
  p = 0.3,
  m = 5,
  maxit = 5,
  printFlag = FALSE,
  ...
)
}
\arguments{
\item{train.data}{A data.frame with missing values}

\item{p}{The proportion of extra missing values}

\item{m}{The number of overimputed datasets. Default: 5}

\item{maxit}{The number of imputation iterations. Default: 1}

\item{...}{Extra arguments to pass to mice}
}
\value{
an overimpute object
}
\description{
Create extra missing values and overimpute them
}
