% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimpute_midae.R
\name{overimpute_midae}
\alias{overimpute_midae}
\title{Overimputation using midae}
\usage{
overimpute_midae(
  train.data,
  test.data = NULL,
  p = 0.2,
  seed = NULL,
  m = 5,
  categorical.encoding = "onehot",
  device = "cpu",
  epochs = 100,
  batch.size = 512,
  subsample = 1,
  early.stopping.epochs = 1,
  dae.params = list(),
  pmm.type = NULL,
  pmm.k = 5,
  pmm.link = "prob",
  pmm.save.vars = NULL,
  loss.na.scale = FALSE,
  verbose = TRUE,
  print.every.n = 1,
  save.model = TRUE,
  path = NULL
)
}
\value{
an overimpute object
}
\description{
Overimputation using midae
}
