% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_2num_facet}
\alias{plot_2num_facet}
\title{Overlay density plots for multiply imputed values for two numeric variables}
\usage{
plot_2num_facet(
  imputation.lists,
  methods,
  var.x,
  var.y,
  original.data,
  true.data = NULL,
  color.pal = NULL,
  xlim = NULL
)
}
\arguments{
\item{methods}{A list of methods?}

\item{var.x}{The name of the variable on the x axis.}

\item{var.y}{The name of the variable on the y axis.}

\item{original.data}{The original data with missing values.}

\item{true.data}{The true data without missing values. This is generally unknown in practice. If the true data is known (e.g., in cases where it is generated by simulation), it can be specified in this argument. The output will then have an extra panel called \code{MaskedTrue}, which shows values originally observed but intentionally made missing.}

\item{color.pal}{A vector of hex color codes for the observed and m sets of imputed values panels. The vector should be of length \code{m+1}. Default: NULL (use "gray40" for the observed panel, use ggplot2 default colors for other panels.)}

\item{xlim}{the limit of the x axis}

\item{imputation.list}{A list of lists....of \code{m} imputed datasets returned by the \code{mixgb} imputer, or other package.}
}
\value{
Histogram with density plots
}
\description{
Plot overlay density curves of observed values versus m sets of imputed values for a specified numeric variable using \pkg{ggplot2}.
}
\examples{
# obtain m multiply datasets
params <- list(max_depth = 3, subsample = 0.8, nthread = 2)
imputed.data <- mixgb(data = nhanes3, m = 3, xgb.params = params, nrounds = 30)

# plot the multiply imputed values for variable "BMPHEAD"
plot_hist(
  imputation.list = imputed.data, var.name = "BMPHEAD",
  original.data = nhanes3
)
}
