% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R, R/plot_methods0.R
\name{plot_density_facet}
\alias{plot_density_facet}
\title{Overlay density plots for multiply imputed values for a single numeric variable}
\usage{
plot_density_facet(
  imputation.lists,
  methods,
  var.name,
  original.data,
  true.data = NULL,
  color.pal = NULL
)

plot_density_facet(
  imputation.lists,
  methods,
  var.name,
  original.data,
  true.data = NULL,
  color.pal = NULL
)
}
\arguments{
\item{imputation.lists}{A list of lists.... \code{m} imputed datasets returned by the \code{mixgb} imputer, or other package.}

\item{methods}{A vector of length \code{m} containing the names of the imputation methods used to generate the \code{m} imputed datasets. Default: NULL (use "Method 1", "Method 2", etc.)}

\item{var.name}{The name of a numeric variable of interest.}

\item{original.data}{The original data with missing values.}

\item{true.data}{The true data without missing values. This is generally unknown in practice. If the true data is known (e.g., in cases where it is generated by simulation), it can be specified in this argument. The output will then have an extra panel called \code{MaskedTrue}, which shows values originally observed but intentionally made missing.}

\item{color.pal}{A vector of hex color codes for the observed and m sets of imputed values panels. The vector should be of length \code{m+1}. Default: NULL (use "gray40" for the observed panel, use ggplot2 default colors for other panels.)}

\item{xlim}{the limit of the x axis}

\item{imputation.list}{A list of \code{m} imputed datasets returned by the \code{mixgb} imputer, or other package.}
}
\value{
Histogram with density plots

Histogram with density plots
}
\description{
Plot overlay density curves of observed values versus m sets of imputed values for a specified numeric variable using \pkg{ggplot2}.

Plot overlay density curves of observed values versus m sets of imputed values for a specified numeric variable using \pkg{ggplot2}.
}
\examples{
# obtain m multiply datasets
params <- list(max_depth = 3, subsample = 0.8, nthread = 2)
# use precomputed imputed datasets shipped with the package
data(imp_nhanes3_newborn)
imputed.data <- imp_nhanes3_newborn

# plot the multiply imputed values for variable "BMPHEAD"
plot_hist(
  imputation.list = imputed.data, var.name = "BMPHEAD",
  original.data = na_nhanes3_newborn
)
}
