% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs.R
\name{plotcat.trelliscope}
\alias{plotcat.trelliscope}
\title{plot categorical variables using trelliscope}
\usage{
plotcat.trelliscope(obj, train.data, test.data = NULL, type = "dodge")
}
\description{
plot categorical variables using trelliscope
}
