% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs.R
\name{plotcon.trelliscope}
\alias{plotcon.trelliscope}
\title{plot continuous variables using trelliscope}
\usage{
plotcon.trelliscope(obj, train.data, test.data = NULL, type = "ridges")
}
\description{
plot continuous variables using trelliscope
}
