% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess input of `vismi.data.frame()` for plotting}
\usage{
preprocess(data, imp_list, m, imp_idx, vars, integerAsFactor)
}
\arguments{
\item{data}{Original data frame with missing values}

\item{imp_list}{A list of imputed data frames}

\item{m}{A single positive integer specifying the number of imputed datasets to plot}

\item{imp_idx}{A vector of integers specifying the indices of imputed datasets to plot}

\item{vars}{A vector of variable names to include in the plot}

\item{integerAsFactor}{Logical, whether to treat integer variables as factors}
}
\description{
Preprocess input of `vismi.data.frame()` for plotting
}
