% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi.R
\name{print.vismi}
\alias{print.vismi}
\title{print method for vismi objects}
\usage{
\method{print}{vismi}(x, ...)
}
\arguments{
\item{x}{An object of class 'vismi' created by the \code{vismi.data.frame()} function.}

\item{...}{Additional arguments (not used).}
}
\description{
vismi Print method for vismi objects
}
