% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelper.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme}
\alias{element_text}
\alias{element_blank}
\alias{margin}
\alias{unit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:element]{element_blank}}, \code{\link[ggplot2:element]{element_text}}, \code{\link[ggplot2:element]{margin}}, \code{\link[ggplot2]{theme}}}

  \item{grid}{\code{\link[grid]{unit}}}
}}

