% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs.R
\name{trelliscope_cat}
\alias{trelliscope_cat}
\title{Create trelliscope display for overimputations of all categorical variables in a dataset}
\usage{
trelliscope_cat(obj, train.data, test.data = NULL, type = "dodge")
}
\description{
Create trelliscope display for overimputations of all categorical variables in a dataset
}
