% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs.R
\name{trelliscope_con}
\alias{trelliscope_con}
\title{Create trelliscope display for overimputations of all continuous variables in a dataset}
\usage{
trelliscope_con(obj, train.data, test.data = NULL, type = "ridge")
}
\description{
Create trelliscope display for overimputations of all continuous variables in a dataset
}
