% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimp.R
\name{overimp}
\alias{overimp}
\title{Overimpute main function}
\usage{
overimp(
  data,
  m = 5,
  p = 0.2,
  test_ratio = 0,
  method = "mixgb",
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame with missing values}

\item{m}{the number of imputation}

\item{p}{the extra proportion of missing values}

\item{test_ratio}{the proportion of test set. Default is 0, meaning no test set.}

\item{method}{can be one of the following: "mixgb","mice", and more in the future.}

\item{seed}{random seed}

\item{...}{other arguments to be passed into the overimp function}
}
\description{
overimp wrapper function to call different imputation methods
}
