% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimpute.R
\name{overimpute}
\alias{overimpute}
\title{Overimpute main function}
\usage{
overimpute(
  data,
  p = 0.2,
  m = 5,
  test_ratio = 0,
  method = "mixgb",
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame with missing values}

\item{p}{the extra proportion of missing values}

\item{m}{the number of imputation}

\item{test_ratio}{the proportion of test set. Default is 0, meaning no test set.}

\item{method}{can be one of the following: "mixgb", "midae", "mivae", "mice","cart" or "ranger"}

\item{...}{other arguments to be passed into the overimpute function}
}
\description{
overimpute wrapper function to call different imputation methods
}
