% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overimpute.R
\name{overimpute}
\alias{overimpute}
\title{Overimpute main function}
\usage{
overimpute(train.data, test.data = NULL, p = 0.2, m = 5, method, ...)
}
\arguments{
\item{train.data}{a data frame with training set data}

\item{test.data}{a data fram with test set data}

\item{p}{the extra proportion of missing values}

\item{m}{the number of imputation}

\item{method}{can be one of the following: "mixgb", "midae", "mivae", "mice","cart" or "ranger"}

\item{...}{other arguments to be passed into the overimpute function}
}
\description{
overimpute
}
