% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1var.R
\name{plot1D}
\alias{plot1D}
\title{Inspection plot of multiply imputed values for a single variable}
\usage{
plot1D(
  imputation.list,
  var.name,
  original.data,
  true.data = NULL,
  color.pal = NULL,
  num.plot = "hist",
  int.plot = "bar",
  cat.plot = "bar"
)
}
\arguments{
\item{imputation.list}{A list of \code{m} imputed datasets returned by the \code{mixgb} imputer, or other package.}

\item{var.name}{The name of a numeric variable of interest.}

\item{original.data}{The original data with missing values.}

\item{true.data}{The true data without missing values. This is generally unknown in practice. If the true data is known (e.g., in cases where it is generated by simulation), it can be specified in this argument. The output will then have an extra panel called \code{MaskedTrue}, which shows values originally observed but intentionally made missing.}

\item{color.pal}{A vector of hex color codes for the observed and m sets of imputed values panels. The vector should be of length \code{m+1}. Default: NULL (use "gray40" for the observed panel, use ggplot2 default colors for other panels.)}

\item{num.plot}{The type of plot for numeric variables. Options are "hist" (default), "box", or "density".}

\item{int.plot}{The type of plot for integer variables. Options are "bar" (default), "hist", "box", or "density".}

\item{cat.plot}{The type of plot for categorical variables. Options are "bar" (default).}
}
\description{
High-level function to plot observed values versus m sets of imputed values for a specified variable
}
\examples{
# use precomputed imputed datasets shipped with the package
data(imp_nhanes3_newborn)
data(na_nhanes3_newborn)
# plot the multiply imputed values for variable "BMPHEAD"
plot1D(
  imputation.list = imp_nhanes3_newborn, var.name = "BMPHEAD",
  original.data = na_nhanes3_newborn
)
}
