% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess data for plotting}
\usage{
preprocess(data, imp_list, vars, integerAsFactor)
}
\arguments{
\item{data}{Original data frame with missing values}

\item{imp_list}{A list of imputed data frames}

\item{vars}{A vector of variable names to include in the plot}

\item{integerAsFactor}{Logical, whether to treat integer variables as factors}
}
\description{
Preprocess data for plotting
}
