% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi-package.R
\docType{package}
\name{vismi-package}
\alias{vismi-package}
\title{vismi: Visual Diagnostics for Multiple Imputation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive suite of static and interactive visual diagnostics for assessing the quality of multiply-imputed data obtained from packages such as 'mixgb' and 'mice'. The package supports inspection of distributional characteristics, diagnostics based on masking observed values and comparing them with re-imputed values, and convergence diagnostics.
}
\references{
Yongshi Deng, Thomas Lumley. (2026), vismi: Visual Diagnostics for Multiple Imputation,
R package version 0.9.3
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agnesdeng.github.io/vismi/}
  \item Report bugs at \url{https://github.com/agnesdeng/vismi/issues}
}

}
\author{
\strong{Maintainer}: Yongshi Deng \email{agnes.yongshideng@gmail.com} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
\keyword{internal}
