% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi-package.R
\docType{package}
\name{vismi-package}
\alias{vismi-package}
\title{vismi: Visual Diagnostics for Multiple Imputation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive suite of visual diagnostic functions for assessing the quality of multiply-imputed data. 'vismi' provides functions for inspecting distribution characteristics, overimputation and convergence diagnostics. Flexible graphics functions are powered by 'plotly', 'ggplot2', and 'trelliscopejs'.
}
\references{
Yongshi Deng, Thomas Lumley. (2026), vismi: Visual Diagnostics for Multiple Imputation, R package version 0.9.2,
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agnesdeng.github.io/vismi/}
  \item Report bugs at \url{https://github.com/agnesdeng/vismi/issues}
}

}
\author{
\strong{Maintainer}: Yongshi Deng \email{agnes.yongshideng@gmail.com} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
\keyword{internal}
