% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi-package.R
\docType{package}
\name{vismi-package}
\alias{vismi-package}
\title{\pkg{vismi}: Visualisation Tools for Multiple Imputation}
\description{
Visual diagnostics for multiply-imputed data
}
\references{
Deng, Y., & Lumley, T. (2021). Multiple Imputation Through XGBoost.
arXiv:2106.01574.
}
