% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi-package.R, R/vismi-package0.R
\docType{package}
\name{vismi-package}
\alias{vismi}
\alias{vismi-package}
\title{vismi: Visualisation Tools for Multiple Imputation}
\description{
Visual diagnostics for multiply-imputed data

Visual diagnostics for multiply-imputed data
}
\references{
Deng, Y., & Lumley, T. (2021). Multiple Imputation Through XGBoost.
arXiv:2106.01574.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agnesdeng/vismi}
  \item \url{https://agnesdeng.github.io/vismi/}
  \item Report bugs at \url{https://github.com/agnesdeng/vismi/issues}
}

}
\author{
\strong{Maintainer}: Yongshi Deng \email{yongshi.deng@auckland.ac.nz} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
\keyword{internal}
