% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi-package.R
\docType{package}
\name{vismi-package}
\alias{vismi-package}
\title{vismi: Visualisation Tools for Multiple Imputation}
\description{
Visual diagnostic tools for assessing multiply imputed datasets created with
'mixgb' or other imputers.
}
\references{
Deng, Y., & Lumley, T. (2026), vismi: Visualisation Tools for Multiple Imputation, arXiv preprint.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agnesdeng.github.io/vismi/}
  \item Report bugs at \url{https://github.com/agnesdeng/vismi/issues}
}

}
\author{
\strong{Maintainer}: Yongshi Deng \email{agnes.yongshideng@gmail.com} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
\keyword{internal}
