% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi_data_frame.R, R/vismi_overimp.R
\name{vismi}
\alias{vismi}
\alias{vismi.data.frame}
\alias{vismi.overimp}
\title{Visualise Multiple Imputations}
\usage{
vismi(...)

\method{vismi}{data.frame}(
  data,
  imp_list,
  x = NULL,
  y = NULL,
  z = NULL,
  interactive = TRUE,
  integerAsFactor = FALSE,
  color_pal = NULL,
  marginal_x = NULL,
  marginal_y = NULL,
  ...
)

\method{vismi}{overimp}(
  obj,
  x = NULL,
  y = NULL,
  z = NULL,
  integerAsFactor = FALSE,
  num_plot = "qq",
  fac_plot = "bar",
  train_color_pal = NULL,
  test_color_pal = NULL,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying plotting functions, such as point_size, alpha, nbins, width, and boxpoints.}

\item{data}{A data frame containing the original data with missing values.}

\item{imp_list}{A list of imputed data frames.}

\item{x}{A character string specifying the name of the variable to plot on the x}

\item{y}{A character string specifying the name of the variable to plot on the y}

\item{z}{A character string specifying the name of the variable to plot on the z}

\item{interactive}{A logical value indicating whether to create an interactive plotly plot (TRUE
by default) or a static ggplot2 plot (FALSE).}

\item{integerAsFactor}{A logical value indicating whether to treat integer variables as factors
(TRUE) or numeric (FALSE). Default is FALSE.}

\item{color_pal}{A named vector of colors for different imputation sets. If NULL
(default), a default color palette is used.}

\item{marginal_x}{A character string specifying the type of marginal plot to add
for the x variable in 2D plots. Options are "hist", "box", "rug", "box+rug", or NULL
(default, no marginal plot) when interactive = TRUE. Options are "box", "rug", "box+rug", or NULL
(default, no marginal plot) when interactive = FALSE.}

\item{marginal_y}{A character string specifying the type of marginal plot to add
for the y variable in 2D plots. Options are "hist", "box", "rug", "box+rug", or NULL
(default, no marginal plot) when interactive = TRUE. Options are "box", "rug", "box+rug", or NULL
(default, no marginal plot) when interactive = FALSE.}

\item{obj}{Overimputation object of class 'overimp' created by the \code{overimpute()} function.}

\item{num_plot}{A character string specifying the type of plot for numeric variables. Options
are "qq" (default) for QQ plot.}

\item{fac_plot}{A character string specifying the type of plot for categorical variables. Options
are "bar" (default) for bar plot.}

\item{int_plot}{A character string specifying the type of plot for integer variables. Options
are "ridge" (default) for ridge plot.}
}
\value{
A plotly or ggplot2 object visualizing the imputed data.
}
\description{
This function provides visual diagnostic tools for assessing multiply imputed datasets created with 'mixgb' or other imputers. It supports 1D, 2D, and 3D visualizations for numeric variables.
}
\section{Methods (by class)}{
\itemize{
\item \code{vismi(data.frame)}: Visualise Multiple Imputations for Distributional Characteristics

\item \code{vismi(overimp)}: Visualise Multiple Imputation via Overimputations

}}
