% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismi_converge.R
\name{vismi_converge}
\alias{vismi_converge}
\alias{vismi_converge.mixgb}
\alias{vismi_converge.mids}
\title{Visualise onvergence diagnostics}
\usage{
vismi_converge(obj, var, tick_vals = NULL, color_pal = NULL)

\method{vismi_converge}{mixgb}(obj, var, tick_vals = NULL, color_pal = NULL)

\method{vismi_converge}{mids}(obj, var, tick_vals = NULL, color_pal = NULL)
}
\arguments{
\item{obj}{A 'mixgb' object returned by \code{mixgb()} function or a 'mids' object returned by the \code{mice()} function.}

\item{var}{The name of the variable to plot convergence for.}

\item{tick_vals}{Optional numeric vector specifying x-axis tick values for iterations.}

\item{color_pal}{A vector of m color codes (e.g., hex codes). If NULL, default colors will be used.}
}
\value{
A ggplot2 object showing the convergence plot for the specified variable.
}
\description{
Visualise onvergence diagnostics

Visualise onvergence diagnostics for a variable imputed by mixgb

Visualise convergence diagnostics for a variable imputed by MICE
}
